Refactor: types out of core into own package
