+++ .changeset/itchy-cameras-prove.md
@@ -1,0 +1,9 @@
+---
+"@stackwright/nextjs": patch
+"@stackwright/themes": patch
+"@stackwright/types": patch
+"@stackwright/core": patch
+"@stackwright/cli": patch
+---
+
+Refactor types out of core into own package.
