+++ packages/cli/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "es2020",
+    "target": "es2022",
     "module": "commonjs",
     "lib": ["es2020"],
     "outDir": "./dist",
