+++ packages/core/src/hooks/useSiteConfig.ts
@@ -1,5 +1,5 @@
 import { createContext, useContext } from 'react';
-import { SiteConfig } from '../types/siteConfig';
+import { SiteConfig } from '../../../types/src/types/siteConfig';
 
 // Create the context
 const SiteConfigContext = createContext<SiteConfig | null>(null);
