--- packages/core/src/types/index.ts
@@ -1,2 +1,0 @@
-export * from './content';
-export * from './siteConfig';
