+++ packages/core/test/components/DynamicPage.snapshot.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import  DynamicPage from '../../src/components/DynamicPage';
-import { ButtonVariant, TypographyVariant } from 'index';
+import { ButtonVariant, TypographyVariant } from '@stackwright/types';
 
 describe('DynamicPage Snapshots', () => {
   it('matches the snapshot', () => {
