+++ packages/core/test/components/DynamicPage.test.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import DynamicPage  from '../../src/components/DynamicPage';
-import { ButtonVariant, TypographyVariant } from 'index';
+import { ButtonVariant, TypographyVariant } from '@stackwright/types';
 
 describe('DynamicPage', () => {
   it('should render the page based on YAML configuration', () => {
