+++ packages/core/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: ['src/index.ts', 'src/components/DynamicPage.ts', 'src/pages/SlugPage.ts'],
   format: ['cjs', 'esm'],
-  target: 'es2020',
+  target: 'es2022',
   dts: true, // TypeScript declarations handled separately
   splitting: false,
   sourcemap: true,
