+++ packages/types/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@stackwright/types",
+  "version": "0.1.0",
+  "type": "module",
+  "description": "TypeScript types and JSON schemas for Stackwright",
+  "main": "./dist/index.js",
+  "module": "./dist/index.mjs", 
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    },
+    "./schemas": {
+      "types": "./dist/schemas/index.d.ts",
+      "import": "./dist/schemas/index.mjs",
+      "require": "./dist/schemas/index.js"
+    },
+    "./schemas/content-schema.json": "./dist/schemas/content-schema.json"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "pnpm generate-schemas && tsup",
+    "generate-schemas": "tsx src/generate-schemas.ts",
+    "dev": "pnpm generate-schemas && tsup --watch",
+    "test": "vitest"
+  },
+  "dependencies": {
+    "@stackwright/themes": "workspace:*",
+    "js-yaml": "^4.1.0"
+  },
+  "devDependencies": {
+    "@types/js-yaml": "^4.0.5",
+    "@types/fs-extra": "11.0.4",
+    "fs-extra": "^11.3.0",
+    "tsup": "^8.0.0",
+    "tsx": "^4.0.0",
+    "typescript": "^5.6.0",
+    "vitest": "^3.2.4",
+    "typescript-json-schema": "^0.65.1"
+  }
+}
