+++ packages/types/src/types/enums.ts
@@ -1,0 +1,9 @@
+export enum GraphicPosition {
+    LEFT = 'left',
+    RIGHT = 'right'
+}
+
+export type TypographyVariant = 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'subtitle1' | 'subtitle2' | 'body1' | 'body2' | 'caption' | 'button' | 'overline'
+export type ButtonVariant = 'text' | 'outlined' | 'contained'
+export type AlignmentVariant = 'left' | 'center' | 'right'
+export type GraphicVariant = 'contained' | 'overflow'
