+++ packages/types/src/types/yaml.d.ts
