+++ packages/types/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "module": "es2022",
     "target": "es2022",
     "lib": ["dom", "dom.iterable", "es6"],
     "allowJs": true,
@@ -13,7 +14,6 @@
     "isolatedModules": true,
     "noEmit": false,
     "jsx": "react-jsx",
-    "module": "commonjs",
     "outDir": "./dist",
     "rootDir": "./src",
     "declaration": true,
